+++ core/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v0.9.0 (YYYY-MM-DD)
 
 - **Deprecation:** Mark method `config.WithCaptureHTTPResponse` as deprecated, to avoid confusion due to it not being a configuration option. Use `runtime.WithCaptureHTTPResponse` instead.
-- **Deprecation:** Mark method `config.WithJWKSEndpoint` as deprecated. Validation using JWKS was removed, for being redundant with token validation done in the APIs. This option has no effect.
+- **Deprecation:** Mark method `config.WithJWKSEndpoint` and field `config.Configuration.JWKSCustomUrl` as deprecated. Validation using JWKS was removed, for being redundant with token validation done in the APIs. These have no effect.
 - **Breaking Change:** Remove method `KeyFlow.Clone`, that was no longer being used.
 
 ## v0.8.0 (2024-02-16)